separate stateless components from containers
